add objects and functions to the script
add three objects for the Gameboard, Players and Gameflow.
Gameboard: First, it generates the board in a 2D Array, which uses the cell function behind it. Added a temporary printBoard for orientation. Then I added a function to markFields with either 1 (Player1) or 4(Playe
r2). Lastly, I added a function to check the board if someone has three in a line, which is calles winConditions.
Gameflow: Added a function, which keeps track of the rounds left to play (empty fields)